Implemented Responsive Images in Bio Component
This commit introduces the functionality to display different images based on the device's screen size in the Bio component. The images object from @data/bioData was used, which contains the URLs for the desktop and mobile versions of the images. The useState and useEffect hooks from React were utilized to listen for changes in the window's width and update the displayed image accordingly. Now, if the screen width is less than 768 pixels, the mobile image will be displayed. Otherwise, the desktop image will be used. This enhancement improves the user experience by ensuring that the appropriate image is displayed for the user's device.